Add replay-first corpus snapshot harness

+++ corpus/c/type/case-01/src/main.c
@@ -1,2 +1,6 @@
 int takes_int(int value) { return value; }
-int main(void) { const char *value = "x"; return takes_int((int)value); }
+
+int main(void) {
+    const char *value = "x";
+    return takes_int(value);
+}
